requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.34.0
tokenizers>=0.12.1,<0.15
sentencepiece==0.1.99
shortuuid
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.41.0
pydantic>=1,<2
markdown2[all]
numpy
scikit-learn==1.2.2
gradio==3.35.2
gradio_client==0.2.9
requests
httpx==0.24.0
uvicorn
fastapi
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
supervision==0.10.0
loguru
python-dotenv